# Use two-step Softmax to do cuda sampling