--- a/src/cuda/cuda_sampling.cuh
+++ b/src/cuda/cuda_sampling.cuh
@@ -12,8 +12,11 @@ struct SamplingData {
   SamplingData(unsigned long long random_seed, int batch_size, int vocab_size, cudaStream_t stream);
   cuda_unique_ptr<int> indices_sorted;
   cuda_unique_ptr<float> scores_sorted;
-  cuda_unique_ptr<float> scores_buffer; // Temporary buffer for sorting/softmaxing scores before sampling
+  cuda_unique_ptr<float> scores_buffer;
   cuda_unique_ptr<float> prefix_sums;
+  cuda_unique_ptr<float> scores_temp;
+  cuda_unique_ptr<float> scores_adjusted;
+  cuda_unique_ptr<float> prefix_sums_adjusted;
   cuda_unique_ptr<float> thresholds;
   cuda_unique_ptr<int> indices_in;
   cuda_unique_ptr<int> offsets;
@@ -26,7 +29,7 @@ void LaunchPopulateIndices(int* indices, int size, int batch_size, cudaStream_t 
 void GetSample(SamplingData* data, cudaStream_t stream, int32_t* d_next_token, float* d_scores, int vocab_size, int batch_size, int k, float p, float temperature);
 
 template <bool is_log_softmax>
-void DispatchBlockwiseSoftmaxForward(cudaStream_t stream, float* output, const float* input, int softmax_elements, int input_stride, int output_stride, int batch_count, float temperature = 1.0);
+void DispatchBlockwiseSoftmaxForward(cudaStream_t stream, float* output, const float* input, int softmax_elements, int input_stride, int output_stride, int batch_count);
 
 }  // namespace cuda
 }  // namespace Generators