--- a/src/cuda/search_cuda.h
+++ b/src/cuda/search_cuda.h
@@ -48,7 +48,7 @@ struct GreedySearch_Cuda : Search_Cuda {
   DeviceSpan<int32_t> GetNextIndices() override { return {}; }
 
   void SelectTop() override { SampleTopKTopP(1, 0.0, 1.0); }
-  void SampleTopK(int k, float t) override { SampleTopKTopP(k, 0.0, t); }
+  void SampleTopK(int k, float t) override { SampleTopKTopP(k, 1.0, t); }
   void SampleTopP(float p, float t) override { SampleTopKTopP(-1, p, t); }
   void SampleTopKTopP(int k, float p, float t) override;
   void AppendTokens(DeviceSpan<int32_t>& next_tokens) override;  // shape (batch_size, sequence_length)